refactor(BEB-8): organiza arquivos de requisicoes e respostas
