--- src/modules/users/dtos/index.ts
@@ -1,3 +1,0 @@
-export * from './auth-user.dto';
-export * from './create-user.dto';
-export * from './auth-user.dto';
