+++ src/modules/users/input/auth-user.input.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsEmail, IsNotEmpty } from 'class-validator';
 
-export class AuthUserDto {
+export class AuthUserInput {
   @ApiProperty({
     name: 'email',
     example: 'anakin@tatooine.com',
@@ -16,18 +16,4 @@ export class AuthUserDto {
   })
   @IsNotEmpty({ message: 'Senha é obrigatória' })
   password: string;
-}
-
-export class AuthUserResponseDto {
-  @ApiProperty({
-    name: 'message',
-    example: 'Login efetuado com sucesso',
-  })
-  message: string;
-  @ApiProperty({
-    name: 'access_token',
-    example:
-      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1MzI1NDI2NC0zMmEwLTRhYWYtODIxZC02MmU3ZTZkNDRhZTUiLCJyb2xlIjoidXNlciIsImlhdCI6MTcxNTIxOTk5Nn0.XpEXFOBfS3p7bBWO7iC5qrcwQQ5jvTp1yE6hUZcDkiw',
-  })
-  access_token: string;
 }
