+++ src/modules/users/input/index.ts
@@ -1,0 +1,2 @@
+export * from './auth-user.input';
+export * from './create-user.input';
