+++ src/modules/users/output/auth-user.output.ts
@@ -1,0 +1,15 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class AuthUserOutput {
+  @ApiProperty({
+    name: 'message',
+    example: 'Login efetuado com sucesso',
+  })
+  message: string;
+  @ApiProperty({
+    name: 'access_token',
+    example:
+      'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1MzI1NDI2NC0zMmEwLTRhYWYtODIxZC02MmU3ZTZkNDRhZTUiLCJyb2xlIjoidXNlciIsImlhdCI6MTcxNTIxOTk5Nn0.XpEXFOBfS3p7bBWO7iC5qrcwQQ5jvTp1yE6hUZcDkiw',
+  })
+  access_token: string;
+}
