+++ src/modules/users/output/index.ts
@@ -1,0 +1,2 @@
+export * from './auth-user.output';
+export * from './user-account.output';
