+++ src/modules/users/output/user-account.output.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Expose } from 'class-transformer';
 
-export class UserAccountDto {
+export class UserAccountOutput {
   @ApiProperty({ name: 'id', example: '5aaca898-fe80-46f7-8530-bbcb837a2f49' })
   @Expose()
   id: string;
